requirements.txt
flask
pandas
scikit-learn
google-generativeai
python-dotenv
